+++ Sources/CGPointExtensions.swift
@@ -10,6 +10,16 @@ import UIKit
 import Foundation
 
 extension CGPoint {
+    
+    /// EZSE: Constructor from CGVector
+    public init(vector: CGVector) {
+        self.init(x: vector.dx, y: vector.dy)
+    }
+    
+    /// EZSE : Constructor from CGFloat
+    public init(angle: CGFloat) {
+        self.init(x: cos(angle), y: sin(angle))
+    }
 
     /// EZSE: Adds two CGPoints.
     public static func + (this: CGPoint, that: CGPoint) -> CGPoint {